Fishnet Grids (Terrain Correction Compartments)
The Fishnet Grids that are used in the computation of terrain corrections. A 25' x 25' geographical grid compartment consisting of 25 grids (each grid is 5' x 5') and the the gravity station is at the compartment center. The other remaining 24 grids around the gravity station are used in the computation of the gravity station terrain correction and thus this is applied to simple bouguer to get complete bouguer anomaly.